login,  profiles

+++ app/Http/Controllers/FotoController.php
@@ -3,8 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Foto;
+use App\Models\TipoFoto;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Storage;
 
 class FotoController extends Controller
 {
@@ -21,6 +23,18 @@ class FotoController extends Controller
         return $foto;
     }
 
+
+    public function fotoTipoOrigen( $idTipoFoto, $origenId )
+    {
+        $foto = Foto::with(['tipoFoto:id,nb_tipo_foto,tx_origen,tx_base_path'])
+                          ->where('id_tipo_foto', $idTipoFoto)
+                          ->where('id_origen', $origenId)
+                          ->get();
+        
+        return $foto;
+    }
+
+
     /**
      * Store a newly created resource in storage.
      *
@@ -30,18 +44,41 @@ class FotoController extends Controller
     public function store(Request $request)
     {
         $validate = request()->validate([
-            'nb_foto'           => 	'required|string|max:100',
-			'tx_src'            => 	'required|string|max:100',
-			'id_tipo_foto'      => 	'required|integer',
-			'id_entidad'        => 	'required|integer',
-			'tx_observaciones'  => 	'nullable|string|max:100',
-			'id_status'         => 	'required|integer',
-			'id_usuario'        => 	'required|integer',
+			'tx_src'       =>  'required|string',
+			'id_tipo_foto' =>  'required|integer',
+			'id_origen'    =>  'required|integer',
+			'id_usuario'   =>  'required|integer',
+        ]);
+
+        $tipoFoto    = TipoFoto::where('id', $request->id_tipo_foto)->first();
+
+        $imgName     = date('U') . '_' . $request->id_origen . '.jpg';
+        $imgSource   = $request->tx_src;
+        $storage     = $tipoFoto->tx_storage;
+        $folder      = $request->id_origen;
+        
+        $stored      = $this->storeImage($imgSource, $storage, $imgName, $folder);
+
+        $request->merge([
+            'tx_src'       =>  $imgName,
+            'nb_foto'      =>  'imagen del Colegio ' . $request->id_origen,
+            'id_status'    =>  1
         ]);
 
         $foto = foto::create($request->all());
 
-        return [ 'msj' => 'Foto Agregado Correctamente', compact('foto') ];
+        return [ 'msj' => 'Foto Agregada Correctamente', compact('foto', 'stored') ];
+    }
+
+    public function storeImage($imgSource, $storage, $imgName, $folder)
+	{
+        $imgSource = substr($imgSource, strpos($imgSource, 'base64,') + 7);
+             
+        $imgSource  = base64_decode($imgSource);
+                                                    
+        $stored = Storage::disk($storage)->put( 'foto' . DIRECTORY_SEPARATOR . $folder . DIRECTORY_SEPARATOR . $imgName, $imgSource);
+
+        return $stored;
     }
 
     /**
@@ -68,7 +105,7 @@ class FotoController extends Controller
             'nb_foto'           => 	'required|string|max:100',
 			'tx_src'            => 	'required|string|max:100',
 			'id_tipo_foto'      => 	'required|integer',
-			'id_entidad'        => 	'required|integer',
+			'id_origen'        => 	'required|integer',
 			'tx_observaciones'  => 	'nullable|string|max:100',
 			'id_status'         => 	'required|integer',
 			'id_usuario'        => 	'required|integer',
@@ -87,8 +124,27 @@ class FotoController extends Controller
      */
     public function destroy(Foto $foto)
     {
-        $foto = $foto->delete();
- 
-        return [ 'msj' => 'Foto Eliminado' , compact('foto')];
+
+        $tipoFoto   = TipoFoto::where('id', $foto->id_tipo_foto)->first();
+
+        $storage    = $tipoFoto->tx_storage;
+
+        $path       = 'foto' . DIRECTORY_SEPARATOR . $foto->id_origen . DIRECTORY_SEPARATOR . $foto->tx_src;
+
+        if($this->deleteImage( $storage, $path ))
+        {
+            $foto = $foto->delete();
+
+            return [ 'msj' => 'Foto Eliminada' , compact('foto')];
+
+        }else
+        {
+            return response('Error al Eliminar la Foto', 403) ;
+        }
+    }
+
+    public function deleteImage( $storage, $path )
+    {       
+        return Storage::disk($storage)->delete($path);
     }
 }
